+++ public/index.html
@@ -3,12 +3,6 @@
 <head>
     <title>RPG Init</title>
     <meta http-equiv="content-type" content="text/html; charset=utf-8" />
-    <link rel="stylesheet" type="text/css" href="css/RPG_init.css" inline>
-
-    <script type="application/javascript" src="../libs/react/react.production.min.js" inline></script>
-    <script type="application/javascript" src="../libs/react/react-dom.production.min.js" inline></script>
-    <script type="application/javascript" src="js/creatures.js" inline></script>
-    <script type="application/javascript" src="js/RPG_init.js" inline></script>
 </head>
 <body class="root">
     <div id="io_bar" class="button_bar" style="display:block;"></div>
@@ -28,6 +22,5 @@
     <div id="encounter_card"></div>
     <div id="creature_card" style="display:none;"></div>
     <div id="loading_overlay" class="loading"><p>Loading</p></div>
-    <script type="text/javascript" src="js/components.js" inline></script>
 </body>
 </html>
