+++ scripts/inline.js
@@ -1,15 +1,16 @@
 const { inlineSource } = require('inline-source');
 const fs = require('fs');
 const path = require('path');
-const htmlpath = path.resolve('src/index.html');
+const htmlpath = path.resolve('build/index.html');
 let html;
 
 inlineSource(htmlpath, {
     compress: true,
-    rootpath: path.resolve('src')
+    rootpath: path.resolve('build'),
+    attribute: false
 }).then((html) => {
-    fs.writeFile("RPG_Init.html", html, function(err) {
-        if(err) {
+    fs.writeFile("RPG_Init.html", html, function (err) {
+        if (err) {
             return console.log(err);
         }
         console.log("RPG_Init.html created");
