+++ src/Constants.js
@@ -1,0 +1,10 @@
+class Constants{
+    // Table colors
+    static selected_color = "#4444ff";
+    static turn_color = "yellow";
+    static dead_color = "grey";
+    static creature_table_color = "#eeeeee";
+    static table_header_color = "#dddddd";
+}
+
+export default Constants;
