+++ src/HelperFunctions.js
@@ -1,0 +1,7 @@
+class HelperFunctions{
+    static randInt(min, max){
+        return Math.floor(Math.random() * (max - min + 1)) + min;
+    }
+}
+
+export default HelperFunctions;
