+++ src/events/CreatureChangeEvent.js
@@ -1,0 +1,5 @@
+class CreatureChangeEvent{
+    static event = new Event('creature_change_event');
+}
+
+export default CreatureChangeEvent;
