+++ src/index.css
@@ -1,0 +1,208 @@
+:root{
+  --background-color: #cccccc;
+  --bar-color: grey;
+  --bar-text-color: black;
+  --button-text: black;
+  --header-color: black;
+  --header-text: white;
+  --table-header-color: #dddddd;
+  --card-color: #bbbbbb;
+  --card-header-color: #dddddd;
+  --table-cell-color: #eeeeee;
+}
+
+body {
+  background-color: var(--background-color);
+  height:100%;
+  margin:0;
+  overflow:auto;
+}
+
+.button_bar {
+  background-color: var(--bar-color);
+  margin:0;
+  padding:5px;
+}
+
+.io_bar_button {
+  color: var(--button-text);
+  margin-right: 20px;
+}
+
+.button_bar_text {
+  color: var(--bar-text-color);
+}
+
+.encounter_header {
+  background-color: var(--header-color);
+  color: var(--header-text);
+  margin: 0px;
+  padding: 5px;
+}
+
+.encounter_title {
+  font-size: 32pt;
+  display: inline-block;
+  margin: 0px;
+}
+
+.button_bar_button {
+  color: var(--button-text);
+  margin-right: 10px;
+}
+
+.right_float_button {
+  float: right;
+  margin-right: 0px;
+  font-weight: bold;
+}
+
+table, th, td.creature_table {
+  border: 1px solid black;
+  border-collapse: collapse;
+}
+
+th, td.creature_table {
+  padding: 5px;
+}
+
+th.creature_table {
+  background-color: var(--table-header-color);
+}
+
+.table_container {
+  display: inline-block;
+  max-width: 40%;
+  overflow: auto;
+  padding: 10px;
+}
+
+.clickable {
+  cursor: pointer;
+}
+
+.blank_card {
+  background-color: var(--card-color);
+  border:1px solid black;
+  padding:10px;
+  position: fixed;
+  overflow-y: auto;
+}
+
+.card_container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+  width: 100%;
+}
+
+.card_header{
+  background-color: var(--card-header-color);
+  font-weight: bold;
+  text-align: center;
+  border: 1px solid black;
+  box-sizing: border-box;
+}
+
+.encounter_card_header {
+  margin-bottom: 10px;
+  font-size: 20pt;
+  min-height: 3ex;
+}
+
+.card_table {
+  width: 100%;
+  margin: auto;
+  margin-top: 5px;
+  margin-bottom: 5px;
+  flex-grow: 0;
+}
+
+.card_input {
+  font-size: 20pt;
+  text-align: center;
+}
+
+.card_input_small {
+  font-size: 12pt;
+}
+
+.card_input_left {
+  float: left;
+}
+
+td.card_input {
+  background-color: var(--table-cell-color);
+}
+
+.ability_score {
+  font-size: 12pt;
+  float: left;
+}
+
+textarea.notes {
+  width: 99%;
+}
+
+div.abilities {
+  margin-bottom: -2px;
+  padding-top: 5px;
+  padding-bottom: 5px;
+  overflow: visible;
+  flex: 0;
+}
+
+.abilities {
+  overflow: auto;
+  flex: 1;
+}
+
+textarea.abilities {
+  width: 99%;
+  height: 99%;
+}
+
+p.abilities {
+  background-color: var(--table-cell-color);
+  margin-top: 0px;
+  margin-bottom: 5px;
+  padding: 5px;
+  border: 1px solid black;
+  box-sizing: border-box;
+  white-space: pre-line;
+}
+
+.abilities_container {
+  overflow: auto;
+  flex: 1;
+  background-color: var(--table-cell-color);
+  margin-top: 0px;
+  margin-bottom: 5px;
+  padding: 5px;
+  border: 1px solid black;
+  box-sizing: border-box;
+}
+
+.sticky {
+position: fixed;
+top: 0;
+width: 100%;
+z-index: 1;
+}
+
+.loading {
+  position: fixed;
+  left: 0px;
+  top: 0px;
+  right: 0px;
+  bottom: 0px;
+  background-color: grey;
+  color: white;
+  font-size: 48px;
+  text-align: center;
+  vertical-align: middle;
+  z-index: 2;
+  opacity: .7;
+  display: none;
+}
